moved deploy logic into own deployer pkg and wrote some tests
