--- cmd/deploy/.gitignore
@@ -1,1 +1,0 @@
-deploy
