--- cmd/deploy/Makefile
@@ -1,2 +1,0 @@
-all:
-	go build -o deploy
